(rishah) Fix 473623 -  Fix up data type casting for strict compilers (r+stejohns)

+++ core/PoolObject.cpp
@@ -383,7 +383,7 @@ range_error:
 		{
 			#ifdef AVMPLUS_VERBOSE
 			if (!toplevel || !toplevel->verifyErrorClass())
-				core->console << "class not found: " << m << " index=" << index << "\n";
+				core->console << "class not found: " << m << " index=" << (uint32_t)index << "\n";
 			#endif
 			if (toplevel)
 				toplevel->throwVerifyError(kClassNotFoundError, core->toErrorString(&m));
